Imported the player character into the main scene.
I had to tweak the player control script by explicitly controlling the
animatior within the script, so this will probably break any other
instances of the script.  To import the character, simply drag it into
scene (the import settings should be consistent across the project), add
the camera target script, as well as the player controller script.  Set
the slide value to 0, the jump height to 2, and the extra height to 0.